+++ tests/test_all.nim
@@ -23,4 +23,5 @@ import t_e2e_integration
 import t_code_quality
 import t_memory_safety
 import t_threat_accuracy
+import t_benchmarks
 import test1